ally/requirements.txt
langchain==0.0.354
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
supabase==1.1.1
asyncpg==0.28.0
python-multipart==0.0.6
tiktoken==0.5.1
resend==0.5.2
pyright==1.1.333
pytest==7.4.3
modal==0.55.3994
html5lib==1.1
openai==1.3.9
chainlit==0.7.700
dill==0.3.7
sympy==1.12
plotly==5.18.0
pycoingecko==3.1.0
pinecone-client==2.2.4
google-cloud-aiplatform==1.38.1